2023001941-02 素材与源码/后端/app1/requirements.txt
fastapi==0.95.1
fastapi_jwt_auth==0.5.0
jose==1.0.0
passlib==1.7.4
pydantic==1.10.5
PyJWT==2.6.0
python_jose==3.3.0
redis==4.5.3
SQLAlchemy==2.0.4
starlette==0.26.1
uvicorn==0.22.0
